rest api develop

+++ src/main/java/com/example/timelineservice/timeline/domain/Post.java
@@ -33,7 +33,7 @@ public class Post {
 
     private String content;
 
-    private Integer likeCnt;
+    private Long likeCnt;
 
     private LocalDateTime createdAt;
